Fix MERT trainer Config._owner binding and mixed-precision input dtype
Verified by actually training on a single GPU (B300, NGPU=1) for both
mert_debugmodel and mert_95m, which surfaced two real bugs:

- MERTTrainer had no nested Config class, so Configurable's __init_subclass__
  bound Config._owner to the inherited Trainer, not MERTTrainer -- every
  config_registry factory was silently building a plain Trainer and running
  the generic decoder train loop instead of MERTTrainer's overrides.
- parallelize_mert always wraps the model in fully_shard (even at
  data_parallel_shard_degree=1), so FSDP2's MixedPrecisionPolicy always casts
  parameters to bf16 during forward; the raw-audio input tensor needs an
  explicit matching cast (cast_forward_inputs=False), and the CQT head's own
  buffers are fp32 and need the input cast to match them, not the ambient
  training dtype.

With both fixes, mert_debugmodel trains 20/20 steps with loss decreasing
4.26 -> 3.85 and finite grad norms, and mert_95m (95.5M params, matching its
name) trains 10/10 steps with loss decreasing 7.45 -> 7.07, sane memory
(~9GiB) and MFU (~2.5%).

+++ torchtitan/experiments/mert/cqt.py
@@ -109,7 +109,10 @@ class ConstantQTransform(Module):
             transformer's ``(B, T, C)`` convention).
         """
         pad = self.max_kernel_length // 2
-        x_B1T = F.pad(waveform_BT, (pad, pad), mode="constant").unsqueeze(1)
+        # Always compute in fp32 regardless of the caller's ambient training
+        # dtype: this is a non-differentiable target (the caller wraps it in
+        # ``torch.no_grad()``), and the kernel buffers are fp32.
+        x_B1T = F.pad(waveform_BT.float(), (pad, pad), mode="constant").unsqueeze(1)
         real_BKF = F.conv1d(x_B1T, self.kernel_real, stride=self.hop_length)
         imag_BKF = F.conv1d(x_B1T, self.kernel_imag, stride=self.hop_length)
         magnitude_BKF = torch.sqrt(real_BKF.pow(2) + imag_BKF.pow(2) + 1e-10)
